Move types into types.ts
This is done to separate the type definition from the state store. Which
should make it more clear to see what types we have
