+++ src/lib/components/Canvas.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
 	import {
 		toolState,
-		ToolState,
 		canvasView,
 		mouseEvents,
 		currentSvgElementIndex
 	} from '$lib/stores/stateStore';
+	import { ToolState } from '$lib/types';
 	import { onMount } from 'svelte';
 	import { svgs } from '$lib/stores/svgStore.js';
 
@@ -32,8 +32,8 @@
 	width="100%"
 	height="100%"
 	viewBox={`
-  ${$canvasView.x - $canvasView.width / ((2 * $canvasView.scale) / 100)} 
-  ${$canvasView.y - $canvasView.height / ((2 * $canvasView.scale) / 100)} 
+  ${$canvasView.position.x - $canvasView.width / ((2 * $canvasView.scale) / 100)}
+  ${$canvasView.position.y - $canvasView.height / ((2 * $canvasView.scale) / 100)}
   ${$canvasView.width / ($canvasView.scale / 100)} 
   ${$canvasView.height / ($canvasView.scale / 100)}
 `}
@@ -58,8 +58,8 @@
 
 	<!-- Render background pattern -->
 	<rect
-		x={$canvasView.x - $canvasView.width / ((2 * $canvasView.scale) / 100)}
-		y={$canvasView.y - $canvasView.height / ((2 * $canvasView.scale) / 100)}
+		x={$canvasView.position.x - $canvasView.width / ((2 * $canvasView.scale) / 100)}
+		y={$canvasView.position.y - $canvasView.height / ((2 * $canvasView.scale) / 100)}
 		width="100%"
 		height="100%"
 		fill="url(#pattern-circles)"
