+++ src/lib/components/ColorPicker.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { user, onlineUsers } from '$lib/stores/stateStore';
 	import { colors } from '../color';
-	import type { User, Color } from '$lib/stores/stateStore';
+	import type { User, Color } from '$lib/types';
 
 enum DisabledBy{
 	CurrentUser,
