+++ src/lib/components/tools/Draw.svelte
@@ -3,12 +3,11 @@
 	import Icons from '$lib/icons/MenuIcons.json';
 	import {
 		toolState,
-		ToolState,
 		canvasMousePosition,
 		canvasMouseDown,
 		canvasView
 	} from '$lib/stores/stateStore';
-	import type { CanvasMousePosition } from '$lib/stores/stateStore';
+	import { type CanvasMousePosition, ToolState } from '$lib/types';
 	import type { Svg } from '$lib/stores/svgStore';
 	import { svgs } from '$lib/stores/svgStore';
 
@@ -20,8 +19,8 @@
 	canvasMouseDown.subscribe(stopDraw);
 
 	function mouseToSvgCoordinates(pos: CanvasMousePosition) {
-		const tx = (pos.x - $canvasView.width / 2) / ($canvasView.scale / 100) + $canvasView.x;
-		const ty = (pos.y - $canvasView.height / 2) / ($canvasView.scale / 100) + $canvasView.y;
+		const tx = (pos.x - $canvasView.width / 2) / ($canvasView.scale / 100) + $canvasView.position.x;
+		const ty = (pos.y - $canvasView.height / 2) / ($canvasView.scale / 100) + $canvasView.position.y;
 		return { x: tx, y: ty };
 	}
