+++ src/lib/stores/stateStore.ts
@@ -1,7 +1,7 @@
 // Store to handle global state
 
 import { writable } from 'svelte/store';
-
+import { ToolState, type CanvasMousePosition, type User, type CanvasView } from '$lib/types';
 
 export const serverId = writable('');
 
@@ -9,47 +9,22 @@ export const viewChat = writable(false);
 
 export const settings = writable(false);
 
-export interface CanvasView {
-  x: number;
-  y: number;
-  width: number; // Default values, will be replaced on mount
-  height: number; // Default values, will be replaced on mount
-  scale: number;
-}
-
 export const canvasView = writable<CanvasView>({
-  x: 0,
-  y: 0,
+  position: {
+    x: 0,
+    y: 0
+  },
   width: 0,
   height: 0,
   scale: 100
 });
 
-export enum ToolState { draw, erase, pan, move }
-
 export const toolState = writable<ToolState>(ToolState.pan);
-
-export interface CanvasMousePosition {
-  x: number;
-  y: number;
-}
 
 export const canvasMousePosition = writable<CanvasMousePosition>({
   x: 0,
   y: 0,
 });
-
-export interface Color {
-    name: string;
-    bg: string;
-    border: string;
-
-}
-
-export interface User{
-	name: string;
-	color: Color;
-}
 
 export const onlineUsers = writable<User[]>([]);
